pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "vex"
version = "2.0.0"
description = "VirusTotal IOC Enrichment Tool for SOC/DFIR workflows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Christian Huhn" },
]
keywords = ["virustotal", "ioc", "threat-intelligence", "dfir", "soc", "mitre-attack", "stix"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vex = "vex.main:main"

[project.urls]
Homepage = "https://github.com/christianhuhn/vex"

[tool.setuptools.packages.find]
include = ["vex*"]
